pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "geo-audit-tool"
version = "0.1.0"
description = "Automated GEO (Generative Engine Optimization) audits"
readme = "README.md"
authors = [
    {name = "Neuralis Black", email = "hello@neuralisblack.com"}
]
license = {text = "MIT"}
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP :: Indexing/Search",
]

dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "playwright>=1.40.0",
    "pydantic>=2.5.0",
    "jinja2>=3.1.0",
    "weasyprint>=60.0",
    "streamlit>=1.29.0",
    "python-dotenv>=1.0.0",
    "openai>=1.3.0",
    "redis>=5.0.0",
    "pytest>=7.4.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[project.scripts]
geo-audit = "geo_audit.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "strict"

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
